Add types button
