+++ src/components/atoms/button/button.tsx
@@ -1,15 +1,5 @@
 import styles from './button.module.css';
-
-export enum types {
-  primary = 'primary',
-  secondary = 'secondary'
-}
-
-interface ButtonProps {
-  children: React.ReactNode;
-  onClick: () => void;
-  type: types;
-}
+import { type ButtonProps, types } from './button.types'
 
 const Button: React.FC<ButtonProps> = ({ children, onClick, type }) => {
   const buttonClasses = `${styles.button} ${type === types.primary ? styles.primary : styles.secondary}`;
@@ -22,3 +12,5 @@ const Button: React.FC<ButtonProps> = ({ children, onClick, type }) => {
 };
 
 export default Button;
+
+export { ButtonProps, types }
