+++ src/components/atoms/button/button.types.ts
@@ -1,0 +1,10 @@
+export enum types {
+  primary = 'primary',
+  secondary = 'secondary'
+}
+
+export interface ButtonProps {
+  children: React.ReactNode;
+  onClick: () => void;
+  type: types;
+}
